From preact to react
Big refactor from preact to react. Also, a few things were outdated.

First, we copied everything over.

Then we had to switch how we used styling. Instead of importing the sheet as an object react can just import the sheet and it will use the styling automatically.

Changed images to be loaded from s3 bucket.

Bunch of react router stuff. Was using preact router that used different rules. Unfortunately, react router sort of demands that we include our layout elements (header, footer) in each component, unless we decide to dive into unnecessary complications. There are ways to do it properly, but it was difficult.

Added keys to iterative lists where needed.

Cleaned up a bunch of imports.

General refactoring stuff. Application is back to a similar working state as before. Unfortunately that means few comments in the code. Also, since the project is in a new repository and was copied over (and I didn't think to commit the unchanged from preact stuff) we can't really see what I changed from the original project, but it is pretty irrelevant. This is an effective starting point anyway. Enough rambling, got work to do.